+++ packages/bridge-nest/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.3.10
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.9.0
+
+## 0.3.9
+
+### Patch Changes
+
+- patch test
+
 ## 0.3.8
 
 ### Patch Changes