+++ packages/domain-social/CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/contracts@4.0.2
+
+## 0.3.0
+
+### Minor Changes
+
+- adding user-wish event
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/errors-nest@0.9.0
+  - @volontariapp/bridge-nest@0.3.10
+
+## 0.2.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @volontariapp/bridge-nest@0.3.9
+
 ## 0.2.3
 
 ### Patch Changes